Change <base> to use PUBLIC_URL
