+++ src/util/array-helpers.ts
@@ -1,4 +1,3 @@
-import { Location } from 'history';
 import { compact, fill, isEmpty, times } from 'lodash';
 import { Nullable } from './types';
 
@@ -82,14 +81,6 @@ export const splice = <T>(
   const newList = list.slice();
   newList.splice(...args);
   return newList;
-};
-
-export const splicePath = (
-  path: string | Location<any>,
-  ...args: Parameters<string[]['splice']>
-) => {
-  path = typeof path === 'string' ? path : path.pathname;
-  return splice(path.split('/'), ...args).join('/');
 };
 
 /**
