+++ typings/declarations.d.ts
@@ -3,7 +3,7 @@
 declare namespace NodeJS {
   interface ProcessEnv {
     readonly NODE_ENV: 'development' | 'production' | 'test';
-    readonly PUBLIC_URL: string;
+    readonly PUBLIC_URL?: string;
   }
 }
